Add support for local-only branches
Before this commit, def latest assumed a repository's current branch
was available at the origin. This broke for two reasons: 1) If the
user generated the branch locally, there would be no corresponding
branch at origin; and, 2) If the user set the local branch to track a
remote branch at something other than origin, this was not properly
querying that other remote. This patch takes code provided by Aaron
Wells at:

http://stackoverflow.com/questions/171550/find-out-which-remote-branch-a-local-branch-is-tracking

and cretes new defs tracking_branch and tracking_remote to determine
these values, and modifies def latest to use them. It also renamed
on_branch? to on_branch, is it's not a boolean.